[Refactor] Helper 함수 분리
redis에서 사용하는 helper 함수(binary search, check array empty)를 별로 관리
