+++ backend/src/redis/index.ts
@@ -1,5 +1,5 @@
 import { createClient, RedisClient } from 'redis';
-import { IPostit } from '@customeTypes/board';
+import { findTargetData, isEmpty } from './helper';
 export default class Redis {
 	static instance: Redis;
 	static client: RedisClient;
@@ -83,18 +83,3 @@ export default class Redis {
 }
 
 const INDEX = 'index';
-
-const findTargetData = (storedDataList, targetId) => {
-	let head = 0;
-	let tail = storedDataList.length - 1;
-	let mid = Math.floor((head + tail) / 2);
-	while (head <= tail) {
-		if (Number(storedDataList[mid].id) < Number(targetId)) head = mid + 1;
-		else if (Number(storedDataList[mid].id) > Number(targetId)) tail = mid - 1;
-		else return [storedDataList[mid], mid];
-		mid = Math.floor((head + tail) / 2);
-	}
-	return [undefined, -1];
-};
-
-const isEmpty = (array) => (array.length === 0 ? true : false);
